feat: add support for `embark.config.js`
This commit introduces support for using `embark.config.js` to calculate the
embark configuration object that is otherwise provided via `embark.json`.

If an `embark.config.js` file is present, it will be used over the
`embark.json` file.  The `embark.config.js` module needs to export either an
object or a function that can be asynchronous and has to return or resolve with
an embark configuration object:

```js
// embark.config.js

module.exports = async function () {
  let config = ...; // do lazy calculation of `embarkConfig`;
  return config;
}
```

+++ packages/core/engine/src/index.ts
@@ -121,8 +121,8 @@ export class Engine {
   }
 
   async generateEmbarkJSON() {
-    if (fs.existsSync('embark.json')) {
-      throw new Error(__('embark.json already there. Will not overwrite'));
+    if (fs.existsSync('embark.json') || fs.existsSync('embark.config.js')) {
+      throw new Error(__('embark.json or embark.config.js already exist. Will not overwrite'));
     }
     return fs.writeFile('embark.json', JSON.stringify(constants.defaultEmbarkConfig, null, 2));
   }
